requirements.txt
python-docx
num2words
docx2pdf
pypdf